+++ modules/serial/vehicle/ros/ros_parser.cc
@@ -18,7 +18,9 @@
 #include "modules/serial/vehicle/ros/ros_parser.h"
 
 #include "cyber/logger/log.h"
+#include "modules/serial/vehicle/ros/protocol/misc_fb.h"
 #include "modules/serial/vehicle/ros/protocol/twist_cmd.h"
+#include "modules/serial/vehicle/ros/protocol/twist_fb.h"
 
 namespace apollo {
 namespace serial {
